[ERM-5124] тип свойств, прописанных в DoubleGis.Erm.Platform.Model.Entities.Enums.Infrastructure.PropertiesToEnumMapper, изменен на соответствующие enum'ы
git-tfs-id: [http://uk-tfs02:8080/tfs/DefaultCollection]$/Erm.Core/Dev/Features/EFCF/BL;C25688

+++ Source/ApplicationServices/2Gis.Erm.BL.Operations/Concrete/Old/Integration/ServiceBus/Export/FlowOrders/SerializeDenialReasonHandler.cs
@@ -50,7 +50,7 @@ namespace DoubleGis.Erm.BL.Operations.Concrete.Old.Integration.ServiceBus.Export
                     Name = x.Name,
                     Description = x.Description,
                     ProofLink = x.ProofLink,
-                    TypeCode = x.Type,
+                    TypeCode = (int)x.Type,
                     IsHidden = !x.IsActive,
                 });
         }
